move to dependencies pkg
