+++ cmd/blockchaincmd/add_validator.go
@@ -8,7 +8,6 @@ import (
 	"strings"
 	"time"
 
-
 	"github.com/ava-labs/avalanche-cli/cmd/flags"
 	"github.com/ava-labs/avalanche-cli/pkg/blockchain"
 	"github.com/ava-labs/avalanche-cli/pkg/cobrautils"
