+++ cmd/nodecmd/local.go
@@ -8,6 +8,8 @@ import (
 	"strings"
 	"time"
 
+	"github.com/ava-labs/avalanche-cli/pkg/dependencies"
+
 	"github.com/ava-labs/avalanche-cli/cmd/flags"
 	"github.com/ava-labs/avalanche-cli/pkg/blockchain"
 	"github.com/ava-labs/avalanche-cli/pkg/cobrautils"
@@ -248,7 +250,7 @@ func localStartNode(_ *cobra.Command, args []string) error {
 		latestAvagoPreReleaseVersion = false
 		latestAvagoReleaseVersion = false
 	}
-	avaGoVersionSetting := node.AvalancheGoVersionSettings{
+	avaGoVersionSetting := dependencies.AvalancheGoVersionSettings{
 		UseCustomAvalanchegoVersion:           useCustomAvalanchegoVersion,
 		UseLatestAvalanchegoPreReleaseVersion: latestAvagoPreReleaseVersion,
 		UseLatestAvalanchegoReleaseVersion:    latestAvagoReleaseVersion,
