+++ pkg/node/local.go
@@ -5,6 +5,7 @@ package node
 import (
 	"encoding/hex"
 	"fmt"
+	"github.com/ava-labs/avalanche-cli/pkg/dependencies"
 	"os"
 	"path/filepath"
 	"strings"
@@ -26,14 +27,14 @@ import (
 func setupAvalancheGo(
 	app *application.Avalanche,
 	avalancheGoBinaryPath string,
-	avaGoVersionSetting AvalancheGoVersionSettings,
+	avaGoVersionSetting dependencies.AvalancheGoVersionSettings,
 	network models.Network,
 	printFunc func(msg string, args ...interface{}),
 ) (string, error) {
 	var err error
 	avalancheGoVersion := ""
 	if avalancheGoBinaryPath == "" {
-		avalancheGoVersion, err = GetAvalancheGoVersion(app, avaGoVersionSetting, network)
+		avalancheGoVersion, err = dependencies.GetAvalancheGoVersion(app, avaGoVersionSetting, network)
 		if err != nil {
 			return "", err
 		}
@@ -55,7 +56,7 @@ func StartLocalNode(
 	defaultFlags map[string]interface{},
 	connectionSettings localnet.ConnectionSettings,
 	nodeSettings []localnet.NodeSetting,
-	avaGoVersionSetting AvalancheGoVersionSettings,
+	avaGoVersionSetting dependencies.AvalancheGoVersionSettings,
 	network models.Network,
 ) error {
 	// initializes directories
